% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_btpredict.R
\name{bernoulli_rng}
\alias{bernoulli_rng}
\title{Bernoulli random number distribution}
\usage{
bernoulli_rng(n, p)
}
\arguments{
\item{n}{the size to return}

\item{p}{probability}
}
\value{
a random number 1 or 0
}
\description{
Bernoulli random number distribution
}
