% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc.R
\name{bpc}
\alias{bpc}
\title{Bayesian Paired comparison  regression models in Stan
This is the main function of the package and it samples the posterior distribution of different models using Stan depending on the parameters choice}
\usage{
bpc(
  data,
  player0,
  player1,
  player0_score = NULL,
  player1_score = NULL,
  result_column = NULL,
  model_type = "bradleyterry",
  solve_ties = "random",
  win_score = "higher",
  priors = NULL,
  chains = 4,
  iter = 2000,
  warmup = 1000,
  show_chain_messages = T
)
}
\arguments{
\item{data}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{player0}{A string with name of the column containing the players 0}

\item{player1}{A string with name of the column containing the players 0}

\item{player0_score}{A string with name of the column containing the scores of players 0}

\item{player1_score}{A string with name of the column containing the scores of players 1}

\item{result_column}{A string with name of the column containing the winners. 0 for player 0, 1 for player 1 and 2 for ties}

\item{model_type}{'bradleyterry' (default), 'davidson' for ties}

\item{solve_ties}{A string for the method of handling ties. 'random' for converting ties randomly, 'remove' for removing the tie occurrencess
'none' to ignore ties. The last requires a model to handle ties such as the davidson if there are ties}

\item{win_score}{A string that indicates if a 'higher' score is winner or if a 'lower' score is winner}

\item{priors}{A list with the parameters for the priors. Possible values ''}

\item{chains}{Number of chains passed to Stan sampling. Positive integer, default=4. For more information consult Stan}

\item{iter}{Number of iterations passed to Stan sampling. Positive integer, default =2000. For more information consult Stan}

\item{warmup}{Number of iteration for the warmup passed to Stan sampling. Positive integer, default 1000.  For more information consult Stan}

\item{show_chain_messages}{Hide chain messages from stan}
}
\value{
An object of the class bpc. This object should be used in conjunction with the several auxiliary functions from the package
}
\description{
Bayesian Paired comparison  regression models in Stan
This is the main function of the package and it samples the posterior distribution of different models using Stan depending on the parameters choice
}
\examples{
 bpc(data=citations_agresti,
              player0 = 'journal1',
              player1 = 'journal2',
              player0_score = 'score1',
              player1_score = 'score2',
              model_type='bradleyterry',
              solve_ties='random',
              win_score = 'higher')
}
