% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpcs-package.R
\docType{package}
\name{bpcs-package}
\alias{bpcs-package}
\title{bpcs - A package for Bayesian Paired Comparison analysis with Stan}
\description{
bpcs - A package for Bayesian Paired Comparison analysis with Stan
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
