% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{brasil_soccer_league}
\alias{brasil_soccer_league}
\title{This is a dataset of the first league of the Brazilian soccer championship. It was reduced and translated from the adaduque/Brasileirao_Dataset repository
It contains 1140 matches and 9 Columns
Time: time of the day (24h)
DayWeek: day of the week
Date: date YY-MM-DD
HomeTeam: name of the team playing home
VisitorTeam: name of the team playing visitor
Round: Round number of the championship
Stadium: Name of the stadium where the game was played
ScoreHomeTeam: number of goals for the home team
ScoreVisitorTeam: number of goals for the visitor}
\description{
This is a dataset of the first league of the Brazilian soccer championship. It was reduced and translated from the adaduque/Brasileirao_Dataset repository
It contains 1140 matches and 9 Columns
Time: time of the day (24h)
DayWeek: day of the week
Date: date YY-MM-DD
HomeTeam: name of the team playing home
VisitorTeam: name of the team playing visitor
Round: Round number of the championship
Stadium: Name of the stadium where the game was played
ScoreHomeTeam: number of goals for the home team
ScoreVisitorTeam: number of goals for the visitor
}
\references{
\url{https://github.com/adaoduque/Brasileirao_Dataset}
}
\keyword{data}
