% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_btpredict.R
\name{calculate_p1_win_and_ties}
\alias{calculate_p1_win_and_ties}
\title{Calcutate the probability for a single row}
\usage{
calculate_p1_win_and_ties(i, standata, parameters, n)
}
\arguments{
\item{i}{row number}

\item{standata}{input standata}

\item{parameters}{input sampled parameters}

\item{n}{number of draws}
}
\value{
an array with probability of win1 and probability of tie
}
\description{
Calcutate the probability for a single row
}
