% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{calculate_prob_from_vector}
\alias{calculate_prob_from_vector}
\title{Calculate the probability of occurrence of a in a vector x}
\usage{
calculate_prob_from_vector(x, a)
}
\arguments{
\item{x}{a vector}

\item{a}{a value in x}
}
\value{
a value of the probability of a in vector x
}
\description{
Calculate the probability of occurrence of a in a vector x
}
