% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{check_convergence_diagnostics}
\alias{check_convergence_diagnostics}
\title{Run cmdstan diagnostics for convergence and print the results in the screen
Thin wrapper over cmdstanr cmdstan_diagnose() function}
\usage{
check_convergence_diagnostics(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\description{
Run cmdstan diagnostics for convergence and print the results in the screen
Thin wrapper over cmdstanr cmdstan_diagnose() function
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
check_convergence_diagnostics(m)
}
}
