% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{check_if_there_are_na}
\alias{check_if_there_are_na}
\title{Check for NA in the specfic columns and returns T or F is there is at least 1 NA in those columns}
\usage{
check_if_there_are_na(
  d,
  player0,
  player1,
  player0_score,
  player1_score,
  result_column
)
}
\arguments{
\item{result_column}{}
}
\value{
T (there are NA) or F (no NA)
}
\description{
Check for NA in the specfic columns and returns T or F is there is at least 1 NA in those columns
}
