% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_checks.R
\name{check_numeric_predictor_matrix}
\alias{check_numeric_predictor_matrix}
\title{Check if all values in the predictor matrix are numeric and not NA. Note that TRUE will be cast to 1 and FALSE will be cast to 0}
\usage{
check_numeric_predictor_matrix(predictor_matrix)
}
\arguments{
\item{predictor_matrix}{a predictor matrix generated by the create_predictor_matrix_with_player_lookup_table function}
}
\value{
TRUE (correct) or FALSE (with problems)
}
\description{
Check if all values in the predictor matrix are numeric and not NA. Note that TRUE will be cast to 1 and FALSE will be cast to 0
}
