% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_checks.R
\name{check_predictors_df_contains_all_players}
\alias{check_predictors_df_contains_all_players}
\title{Check if the predictor df contains all players and only those}
\usage{
check_predictors_df_contains_all_players(predictor_df, lookup_table)
}
\arguments{
\item{predictor_df}{the predictor input data frame}

\item{lookup_table}{a lookup table of the players}
}
\value{
TRUE (correct) or FALSE (with problems)
}
\description{
Check if the predictor df contains all players and only those
}
