% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{check_result_column}
\alias{check_result_column}
\title{Check if a data frame column contains only the values 1 0 and 2}
\usage{
check_result_column(d_column)
}
\arguments{
\item{d_column}{}
}
\value{
T (correct) or F (with problems)
}
\description{
Check if a data frame column contains only the values 1 0 and 2
}
