% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_checks.R
\name{check_z_column}
\alias{check_z_column}
\title{Check if a data frame column contains only the values 1 or 0. For the z column}
\usage{
check_z_column(d_column)
}
\arguments{
\item{d_column}{a column of a data frame to be tested}
}
\value{
TRUE (correct) or FALSE (with problems)
}
\description{
Check if a data frame column contains only the values 1 or 0. For the z column
}
