% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{citations_agresti}
\alias{citations_agresti}
\title{This is the citations data from  Agresti (2003) p.448
This data refers to journal citations among four statistics journals during 1987᎐1989. The more often articles in a particular journal are cited, the more prestige that journal accrues.
It contains the how many journal1 was cited by journal2 and journal2 was cited by journal1}
\description{
This is the citations data from  Agresti (2003) p.448
This data refers to journal citations among four statistics journals during 1987᎐1989. The more often articles in a particular journal are cited, the more prestige that journal accrues.
It contains the how many journal1 was cited by journal2 and journal2 was cited by journal1
}
\references{
Agresti, Alan. Categorical data analysis. Vol. 482. John Wiley & Sons, 2003.
}
\keyword{data}
