% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{compute_scores}
\alias{compute_scores}
\title{Giving a player0 an player1 it adds a column to the data frame containing who won or if it was a tie}
\usage{
compute_scores(
  data,
  player0_score,
  player1_score,
  solve_ties = "random",
  win_score = "higher"
)
}
\arguments{
\item{player0_score}{name of the column in data}

\item{player1_score}{name of the column in data}

\item{solve_ties}{Method to solve the ties, either randomly allocate, or do nothing, or remove the row from the datasetc('random', 'none', 'remove').}

\item{win_score}{decides if who wins is the one that has the highest score or the lowest score}
}
\value{
a dataframe with column 'y' that contains the results of the comparison
}
\description{
Giving a player0 an player1 it adds a column to the data frame containing who won or if it was a tie
}
