% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{compute_ties}
\alias{compute_ties}
\title{Giving a result column we create a new column with ties (0 and 1 if it has)}
\usage{
compute_ties(d, result_column)
}
\arguments{
\item{d}{data frame}

\item{result_column}{column where the result is}
}
\value{
dataframe with a column called ties
}
\description{
Giving a result column we create a new column with ties (0 and 1 if it has)
}
