% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{create_array_of_par_names}
\alias{create_array_of_par_names}
\title{Create an array with the name of the parameter and the corresponding index
lambda\link{1} --> lambda_Biometrika}
\usage{
create_array_of_par_names(par, lookup_table)
}
\arguments{
\item{par}{a name for the parameter}

\item{lookup_table}{a lookuptable to convert the indexes to that}
}
\value{
a vector
}
\description{
Create an array with the name of the parameter and the corresponding index
lambda\link{1} --> lambda_Biometrika
}
