% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_cluster_index_with_existing_lookup_table}
\alias{create_cluster_index_with_existing_lookup_table}
\title{Create two columns with the indexes for the names
Here we use an existing lookup table. Should be used in predicting}
\usage{
create_cluster_index_with_existing_lookup_table(
  d,
  cluster,
  cluster_lookup_table
)
}
\arguments{
\item{d}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{cluster}{The name of the column of data data contains player0}

\item{cluster_lookup_table}{a lookup table for the cluster}
}
\value{
A dataframe with the additional columns 'player0_index' and 'player1_index' that contains the indexes
}
\description{
Create two columns with the indexes for the names
Here we use an existing lookup table. Should be used in predicting
}
