% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_index}
\alias{create_index}
\title{Create two columns with the indexes for the names of the players
Here we create a new lookup table. Should be used when sampling the parameters}
\usage{
create_index(d, player0, player1)
}
\arguments{
\item{d}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{player0}{The name of the column of data data contains player0}

\item{player1}{The name of the column of data data contains player0}
}
\value{
A dataframe with the additional columns 'player0_index' and 'player1_index' that contains the indexes
}
\description{
Create two columns with the indexes for the names of the players
Here we create a new lookup table. Should be used when sampling the parameters
}
