% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_index_cluster_lookuptable}
\alias{create_index_cluster_lookuptable}
\title{Create a lookup table of names and indexes
Note that the indexes will be created in the order they appear. For string this does not make much difference but for numbers the index might be different than the actual number that appears in names}
\usage{
create_index_cluster_lookuptable(d, cluster)
}
\arguments{
\item{d}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{cluster}{A string with the name of the cluster variable}
}
\value{
A dataframe of a lookup table with columns Names and Index
}
\description{
Create a lookup table of names and indexes
Note that the indexes will be created in the order they appear. For string this does not make much difference but for numbers the index might be different than the actual number that appears in names
}
