% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{create_index_lookuptable}
\alias{create_index_lookuptable}
\title{Create a lookup table of names and indexes}
\usage{
create_index_lookuptable(data, player0, player1)
}
\arguments{
\item{data}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{player0}{The name of the column of data data contains player0}

\item{player1}{The name of the column of data data contains player0}
}
\value{
A dataframe of a lookup table
}
\description{
Create a lookup table of names and indexes
}
