% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_index_predictors_with_lookup_table}
\alias{create_index_predictors_with_lookup_table}
\title{Receives one column with player names and returns a data frame with the relevant index columns based on a given lookup table
To be used with the predictors data frame}
\usage{
create_index_predictors_with_lookup_table(d, player, lookup_table)
}
\arguments{
\item{d}{a data frame of the predictors}

\item{player}{The name of the column of data data contains the player}

\item{lookup_table}{a lookup table data frame}
}
\value{
A dataframe with the additional column 'player_index'
}
\description{
Receives one column with player names and returns a data frame with the relevant index columns based on a given lookup table
To be used with the predictors data frame
}
