% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_index_with_existing_lookup_table}
\alias{create_index_with_existing_lookup_table}
\title{Create two columns with the indexes for the names
Here we use an existing lookup table. Should be used in predicting}
\usage{
create_index_with_existing_lookup_table(d, player0, player1, lookup_table)
}
\arguments{
\item{d}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{player0}{The name of the column of data data contains player0}

\item{player1}{The name of the column of data data contains player0}

\item{lookup_table}{lookup_table a lookup table data frame}
}
\value{
A dataframe with the additional columns 'player0_index' and 'player1_index' that contains the indexes
}
\description{
Create two columns with the indexes for the names
Here we use an existing lookup table. Should be used in predicting
}
