% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_predictor_matrix_with_player_lookup_table}
\alias{create_predictor_matrix_with_player_lookup_table}
\title{Receives a predictor dataframe, a string with the column of the player, a vector of strings with the columns for the predictors and a lookup table and returns an ordered matrix for Stan
To be used with the predictors data frame}
\usage{
create_predictor_matrix_with_player_lookup_table(
  d,
  player,
  predictors_columns,
  lookup_table
)
}
\arguments{
\item{d}{a data frame of the predictors}

\item{player}{The name of the column of data data contains the player}

\item{predictors_columns}{a vector with strings containing the columns for the predictors}

\item{lookup_table}{a lookup table data frame}
}
\value{
A matrix to be used in stan
}
\description{
Receives a predictor dataframe, a string with the column of the player, a vector of strings with the columns for the predictors and a lookup table and returns an ordered matrix for Stan
To be used with the predictors data frame
}
