% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_predictors_lookup_table}
\alias{create_predictors_lookup_table}
\title{Receives a vector with predictors strings (the column names) and returns a predictor_lookup_table}
\usage{
create_predictors_lookup_table(predictors_columns)
}
\arguments{
\item{predictors_columns}{a vector with strings containing the columns for the predictors}
}
\value{
A matrix to be used in stan
}
\description{
Receives a vector with predictors strings (the column names) and returns a predictor_lookup_table
}
