% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_subject_predictor_lookuptable}
\alias{create_subject_predictor_lookuptable}
\title{Create a lookup table for the subject Predictors}
\usage{
create_subject_predictor_lookuptable(subject_predictors)
}
\arguments{
\item{subject_predictors}{a vectore containing the names of the subject predictors}
}
\value{
a dataframe with a lookup table
}
\description{
Create a lookup table for the subject Predictors
}
