% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_fit}
\alias{get_fit}
\title{Retrieve the cmstanr fit object
This object can be used with any other function or package. You can also convert it to a  stanfit objects from rstan}
\usage{
get_fit(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\value{
a cmdstanr fit object
}
\description{
Retrieve the cmstanr fit object
This object can be used with any other function or package. You can also convert it to a  stanfit objects from rstan
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
fit<- get_fit(m)
print(class(fit))
}
}
