% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_hpdi_parameters}
\alias{get_hpdi_parameters}
\title{Return a data frame with the mean and the HPDI for all parameters}
\usage{
get_hpdi_parameters(bpc_object)
}
\arguments{
\item{bpc_object}{}
}
\value{
a data frame containing a column with the parameters, a column with mean and two columns with higher and lower hpdi
}
\description{
Return a data frame with the mean and the HPDI for all parameters
}
\examples{

}
