% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_loo}
\alias{get_loo}
\title{Tiny wrapper for the PSIS-LOO-CV method from the loo package}
\usage{
get_loo(bpc_object)
}
\arguments{
\item{bpc_object}{}
}
\value{

}
\description{
Tiny wrapper for the PSIS-LOO-CV method from the loo package
}
