% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{get_model_parameters}
\alias{get_model_parameters}
\title{Return all the name of parameters in a model from a bpc_object.
Here we exclude the log_lik and the lp__ since they are not parameters of the model}
\usage{
get_model_parameters(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\value{
a vector with the name of the parameters
}
\description{
Return all the name of parameters in a model from a bpc_object.
Here we exclude the log_lik and the lp__ since they are not parameters of the model
}
