% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Get the location of the output directory for the cmdstanr csv fukes}
\usage{
get_output_dir(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\description{
Get the location of the output directory for the cmdstanr csv fukes
}
