% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_par_names}
\alias{get_par_names}
\title{Create an array with the name of the parameter and the corresponding index
lambda\link{1} --> lambda_Biometrika}
\usage{
get_par_names(bpc_object, par)
}
\arguments{
\item{bpc_object}{is a bpc object}

\item{par}{a name for the parameter}
}
\value{
a vector
}
\description{
Create an array with the name of the parameter and the corresponding index
lambda\link{1} --> lambda_Biometrika
}
