% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_parameters.R
\name{get_parameters_df}
\alias{get_parameters_df}
\title{Function to retrieve a data frame with summary of the parameters}
\usage{
get_parameters_df(
  bpc_object,
  params = NULL,
  HPDI = TRUE,
  credMass = 0.95,
  n_eff = FALSE,
  Rhat = FALSE
)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{params}{a vector with the parameters to use. If null then all will be present}

\item{HPDI}{should return the HPDI or the credible intervals. Default is returning the HPDI}

\item{credMass}{probability mass for the summary stats}

\item{n_eff}{Should include the number of effective samples in the df}

\item{Rhat}{Should include the Rhat in the df}
}
\description{
Function to retrieve a data frame with summary of the parameters
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
param<-get_parameters_df(m)
print(param)
}
}
