% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_publication.R
\name{get_parameters_plot}
\alias{get_parameters_plot}
\title{Return a publication-ready plot for the parameters estimates based on the HPD interval
The returned plot is a caterpillar type of plot}
\usage{
get_parameters_plot(
  bpc_object,
  HPDI = T,
  params = c("lambda"),
  title = "Parameter estimates",
  subtitle = NULL,
  xaxis = "Parameter",
  yaxis = "Value",
  rotate_x_labels = FALSE,
  APA = TRUE,
  keep_par_name = TRUE
)
}
\arguments{
\item{bpc_object}{a bpc fitted object}

\item{HPDI}{use HPD (TRUE) or credible intervals (FALSE) for the plots}

\item{params}{a vector of string for of the parameters to be plotted}

\item{title}{the title of the plot}

\item{subtitle}{optional subtitle for the plot}

\item{xaxis}{title of the x axis}

\item{yaxis}{title of the y axis}

\item{rotate_x_labels}{should the labels be shown horizontally (default, FALSE) or vertically (TRUE)}

\item{APA}{should the graphic be formatted in APA style (default TRUE)}

\item{keep_par_name}{keep the parameter name e.g. lambda\link{Graff} instead of Graff. Default to T. Only valid for lambda, so we can have better ranks}
}
\value{
a ggplot2 caterpillar plot
}
\description{
Return a publication-ready plot for the parameters estimates based on the HPD interval
The returned plot is a caterpillar type of plot
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
p<-get_parameters_plot(m)
p
}
}
