% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_parameters.R
\name{get_parameters_posterior}
\alias{get_parameters_posterior}
\title{Return a dataframe with the posterior distribution of each parameter used in the model}
\usage{
get_parameters_posterior(bpc_object, n = 100)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{n}{number of samples in the posterior}
}
\value{
a dataframe
}
\description{
Return a dataframe with the posterior distribution of each parameter used in the model
}
\examples{
 \donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
param<-get_parameters_posterior(m)
print(head(param))
}
}
