% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_probabilities}
\alias{get_probabilities}
\title{Get the win/draw probabilities based on the strength parameters only}
\usage{
get_probabilities(bpc_object, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{n}{number of samples to draw from the posterior}
}
\value{
a data frame with the respective probabilities
}
\description{
Get the win/draw probabilities based on the strength parameters only
}
