% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_probabilities.R
\name{get_probabilities_newdata}
\alias{get_probabilities_newdata}
\title{Get the empirical win/draw probabilities from a newdata frame. Ths allows the user to specify which specific probabilities are desired}
\usage{
get_probabilities_newdata(bpc_object, newdata, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{newdata}{a data frame equivalent to the one used to fit the model with the specific desired probabilities}

\item{n}{number of samples to draw from the posterior}
}
\value{
a data frame table with the respective probabilities
}
\description{
Get the empirical win/draw probabilities from a newdata frame. Ths allows the user to specify which specific probabilities are desired
}
