% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_probabilities.R
\name{get_probabilities_newdata_df}
\alias{get_probabilities_newdata_df}
\title{Get the empirical win/draw probabilities from a newdata frame. Returning a dataframe with the results
Instead of calculating from the probability formula given from the model we create a predictive posterior distribution for all pair combinations and calculate the posterior wins/loose/draw}
\usage{
get_probabilities_newdata_df(bpc_object, newdata, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{newdata}{a data frame equivalent to the one used to fit the model with the specific desired probabilities}

\item{n}{number of samples to draw from the posterior}
}
\value{
a data frame table with the respective probabilities
}
\description{
Get the empirical win/draw probabilities from a newdata frame. Returning a dataframe with the results
Instead of calculating from the probability formula given from the model we create a predictive posterior distribution for all pair combinations and calculate the posterior wins/loose/draw
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
prob<-get_probabilities_newdata_df(m, newdata=tennis_agresti[c(1,15),])
print(prob)
}
}
