% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_probabilities.R
\name{get_probabilities_newdata_posterior}
\alias{get_probabilities_newdata_posterior}
\title{Get the empirical win/draw probabilities from a newdata frame. Returns a matrix of the posteior}
\usage{
get_probabilities_newdata_posterior(bpc_object, newdata, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{newdata}{a data frame equivalent to the one used to fit the model with the specific desired probabilities}

\item{n}{number of samples to draw from the posterior}
}
\value{
a matrix with the corresponding posterior
}
\description{
Get the empirical win/draw probabilities from a newdata frame. Returns a matrix of the posteior
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
post<-get_probabilities_newdata_posterior(m, newdata=tennis_agresti[c(1,15),])
print(post)
}
}
