% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_probabilities.R
\name{get_probabilities_posterior}
\alias{get_probabilities_posterior}
\title{Get the posterior of the probabilities}
\usage{
get_probabilities_posterior(
  bpc_object,
  newdata = NULL,
  n = 1000,
  model_type = NULL
)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{newdata}{default to NULL. If used,  it will calculate the probabilities only for the newdata. Otherwise it will calculate for all combinations}

\item{n}{number of samples to draw from the posterior}

\item{model_type}{when dealing with some models (such as random effects) one might want to make predictions using the estimated parameters with the random effects but without specifying the specific values of random effects to predict. Therefore one can set a subset of the model to make predictions. For example: a model sampled with bt-U can be used to make predictions of the model bt only.}
}
\value{
a matrix with the corresponding posterior
}
\description{
Get the posterior of the probabilities
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
prob<-get_probabilities_posterior(m)
print(prob)
}
}
