% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_probabilities.R
\name{get_probabilities_table}
\alias{get_probabilities_table}
\title{Publication-ready table for the probabilities}
\usage{
get_probabilities_table(
  bpc_object,
  newdata = NULL,
  n = 100,
  format = "latex",
  digits = 3,
  caption = "Estimated posterior probabilites",
  model_type = NULL
)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{newdata}{default to NULL. If used,  it will calculate the probabilities only for the newdata. Otherwise it will calculate for all combinations}

\item{n}{Number of times to sample from the posterior}

\item{format}{A character string. same formats utilized in the knitr::kable function
\itemize{
\item 'latex': output in latex format
\item 'simple': appropriated for the console
\item 'pipe': Pandoc's pipe tables
\item 'html': for html formats
\item 'rst'
Possible values are latex, html, p, simple (Pandoc's simple tables), and rst.
}}

\item{digits}{number of digits in the table}

\item{caption}{a string containing the caption of the table}

\item{model_type}{when dealing with some models (such as random effects) one might want to make predictions using the estimated parameters with the random effects but without specifying the specific values of random effects to predict. Therefore one can set a subset of the model to make predictions. For example: a model sampled with bt-U can be used to make predictions of the model bt only.}
}
\value{
a formatted table
}
\description{
Publication-ready table for the probabilities
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
t<-get_probabilities_table(m)
print(t)
}
}
