% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_rank_of_players.R
\name{get_rank_of_players}
\alias{get_rank_of_players}
\title{Generate a ranking of the ability based on sampling the posterior distribution of the ranks.
This is not exported. Use either the get_rank_of_players_df or the get_rank_of_players_posterior functions}
\usage{
get_rank_of_players(bpc_object, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{n}{Number of times we will sample the posterior}
}
\value{
a list containing the data frame that represents the table and a matrix containing the posterior distribution of the ranks
}
\description{
Generate a ranking of the ability based on sampling the posterior distribution of the ranks.
This is not exported. Use either the get_rank_of_players_df or the get_rank_of_players_posterior functions
}
