% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_rank_of_players.R
\name{get_rank_of_players_posterior}
\alias{get_rank_of_players_posterior}
\title{Generate a ranking of the ability based on sampling the posterior distribution of the ranks.}
\usage{
get_rank_of_players_posterior(bpc_object, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{n}{Number of times we will sample the posterior}
}
\value{
a matrix containing the posterior distribution of the ranks
}
\description{
To print this object you should remove the last column PosteriorRank since it contain the whole posterior distribution for each case
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
rank_m<-get_rank_of_players_posterior(m,n=100)
print(rank_m)
}
}
