% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_rank_of_players.R
\name{get_rank_of_players_table}
\alias{get_rank_of_players_table}
\title{Publication-ready table for the rank table}
\usage{
get_rank_of_players_table(
  bpc_object,
  format = "latex",
  digits = 3,
  caption = "Estimated posterior ranks",
  n = 1000
)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{format}{A character string. same formats utilized in the knitr::kable function
\itemize{
\item 'latex': output in latex format
\item 'simple': appropriated for the console
\item 'pipe': Pandoc's pipe tables
\item 'html': for html formats
\item 'rst'
\item 'simple' appropriated for the console
}}

\item{digits}{number of digits in the table}

\item{caption}{a string containing the caption of the table}

\item{n}{number of times we are sampling the posterior}
}
\value{
a formatted table
}
\description{
Publication-ready table for the rank table
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
t<-get_rank_of_players_table(m)
print(t)
}
}
