% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_stanfit}
\alias{get_stanfit}
\title{Retrieve the stanfit object generated by rstan.
This object can be used with any other function or package that uses rstan objects}
\usage{
get_stanfit(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\value{
a stanfit object
}
\description{
Retrieve the stanfit object generated by rstan.
This object can be used with any other function or package that uses rstan objects
}
\examples{
m_citations<-bpc(data=citations_agresti,player0 = 'journal1',player1 = 'journal2',player0_score = 'score1',player1_score = 'score2',model_type='bradleyterry',solve_ties='random',win_score = 'higher',show_chain_messages=F)
stanfit<- get_stanfit(m_citations)
print(class(stanfit))
}
