% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_stanfit_summary}
\alias{get_stanfit_summary}
\title{Get stanfit summary table of all parameters excluding log_lik.}
\usage{
get_stanfit_summary(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\value{
a data frame with the summary including quantiles, Rhat and neff
}
\description{
Important to investigate the neff and the Rhat from the MCMC
This excludes the log_lik paramter
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
s <- get_stanfit_summary(m)
print(s)
}
}
