% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_waic}
\alias{get_waic}
\title{iny wrapper for the WAIC method from the loo package}
\usage{
get_waic(bpc_object)
}
\arguments{
\item{bpc_object}{}
}
\value{

}
\description{
iny wrapper for the WAIC method from the loo package
}
