% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_hpd.R
\name{HPD_lower_from_column}
\alias{HPD_lower_from_column}
\title{Returns the lower value of the HPD interval for a data frame column}
\usage{
HPD_lower_from_column(column, credMass = 0.95)
}
\arguments{
\item{column}{the data to calculate the HPDI}

\item{credMass}{Credibility mass for the interval (area contained in the interval)}
}
\value{
the value of the lower HPD interval for that column
}
\description{
Returns the lower value of the HPD interval for a data frame column
}
\references{
Mike Meredith and John Kruschke (2020). HDInterval: Highest (Posterior) Density Intervals. R package version 0.2.2. https://CRAN.R-project.org/package=HDInterval
}
