% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{HPDI_from_stanfit}
\alias{HPDI_from_stanfit}
\title{Calculate HPDI from a stanfit model}
\usage{
HPDI_from_stanfit(stanfit)
}
\arguments{
\item{stanfit}{a stanfit object}
}
\value{
a data frame with the HPDI calculated from coda
}
\description{
Calculate HPDI from a stanfit model
}
