% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit function
See https://en.wikipedia.org/wiki/Logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{is a real -inf to inf}
}
\value{
a value between 0 and 1
}
\description{
Inverse logit function
See https://en.wikipedia.org/wiki/Logit
}
