% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{load_bpc_model}
\alias{load_bpc_model}
\title{Thin wrapper to load a saved bpc model for examining later}
\usage{
load_bpc_model(file_name_with_path)
}
\arguments{
\item{file_name_with_path}{the file name with path following the conventions of the operating system that points to a saved bpc object including the RDS}
}
\description{
Thin wrapper to load a saved bpc model for examining later
}
