% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{Logit function
See https://en.wikipedia.org/wiki/Logit}
\usage{
logit(x)
}
\arguments{
\item{x}{p is a probability 0 to 1}
}
\value{
a value between -inf and inf
}
\description{
Logit function
See https://en.wikipedia.org/wiki/Logit
}
