% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{match_cluster_names_to_cluster_lookup_table}
\alias{match_cluster_names_to_cluster_lookup_table}
\title{Receives a column with cluster names and returns a data frame with the relevant index column based on a given cluster lookup table}
\usage{
match_cluster_names_to_cluster_lookup_table(d, cluster, cluster_lookup_table)
}
\arguments{
\item{d}{a data frame}

\item{cluster}{The name of the column of data data contains player0}

\item{cluster_lookup_table}{a lookup table for the cluster}
}
\value{
A dataframe with the additional columns 'cluster_index' that contains the indexes
}
\description{
Receives a column with cluster names and returns a data frame with the relevant index column based on a given cluster lookup table
}
