% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{match_player_names_to_lookup_table}
\alias{match_player_names_to_lookup_table}
\title{Receives two columns with player names and returns a data frame with the relevant index columns based on a given lookup table}
\usage{
match_player_names_to_lookup_table(d, player0, player1, lookup_table)
}
\arguments{
\item{d}{a data frame}

\item{player0}{The name of the column of data data contains player0}

\item{player1}{The name of the column of data data contains player1}

\item{lookup_table}{a lookup table data frame}
}
\value{
A dataframe with the additional columns 'player0_index' and 'player1_index' that contains the indexes
}
\description{
Receives two columns with player names and returns a data frame with the relevant index columns based on a given lookup table
}
