% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optimization_algorithms}
\alias{optimization_algorithms}
\title{Dataset containing an example of the performance of different optimization algorithms against different benchmark functions.
This is a reduced version of the dataset presented at the paper: "Statistical Models for the Analysis of Optimization Algorithms with Benchmark Functions.".
For details on how the data was collected we refer to the paper.}
\format{
This is the expansion of the data where each row contains 1 match only
\itemize{
\item Algorithm: name of algorithm
\item Benchmark: name of the benchmark problem
\item TrueRewardDifference: Difference between the minimum function value obtained by the algorithm and the known global minimum
\item Ndimensions: Number of dimensions of the benchmark problem
\item MaxFevalPerDimensions: Maximum allowed budget for the algorithm per dimensions of the benchmark problem
\item simNumber: id of the simulation. Indicates the repeated measures of each algorithm in each benchmark
}
}
\source{
Mattos, David Issa, Jan Bosch, and Helena Holmstrom Olsson. Statistical Models for the Analysis of Optimization Algorithms with Benchmark Functions. arXiv preprint arXiv:2010.03783 (2020).
}
\usage{
optimization_algorithms
}
\description{
Dataset containing an example of the performance of different optimization algorithms against different benchmark functions.
This is a reduced version of the dataset presented at the paper: "Statistical Models for the Analysis of Optimization Algorithms with Benchmark Functions.".
For details on how the data was collected we refer to the paper.
}
\keyword{data}
