% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{plot.bpc}
\alias{plot.bpc}
\title{S3 plot function for the parameter plot of a bpc model
This is just a wrapper for the get_parameters_plot function and can be used interchangebly}
\usage{
\method{plot}{bpc}(
  x,
  y = NULL,
  HPDI = T,
  params = c("lambda"),
  title = "Strength estimates",
  subtitle = NULL,
  xaxis = "Player",
  yaxis = "Value",
  rotate_x_labels = FALSE,
  APA = TRUE,
  keep_par_name = FALSE,
  ...
)
}
\arguments{
\item{x}{a bpc object}

\item{y}{Not used. Default to NULL}

\item{HPDI}{use HPD (TRUE) or credible intervals (FALSE) for the plots}

\item{params}{a vector of string for of the parameters to be plotted}

\item{title}{the title of the plot}

\item{subtitle}{optional subtitle for the plot}

\item{xaxis}{title of the x axis}

\item{yaxis}{title of the y axis}

\item{rotate_x_labels}{should the labels be shown horizontally (default, FALSE) or vertically (TRUE)}

\item{APA}{should the graphic be formatted in APA style (default TRUE)}

\item{keep_par_name}{keep the parameter name e.g. lambda Graff instead of Graff. Default to T. Only valid for lambda, so we can have better ranks}

\item{...}{additional parameters for the generic S3 plot function. Not used.}
}
\value{
a ggplot2 caterpillar plot
}
\description{
S3 plot function for the parameter plot of a bpc model
This is just a wrapper for the get_parameters_plot function and can be used interchangebly
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
p<-get_parameters_plot(m)
p
}
}
