% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{posterior_predictive}
\alias{posterior_predictive}
\title{Calculate the posterior predictive distributions.
Helps to check the fitness of the model. Use it in conjunction with the shinystan app to visualize some nice posterior predictive plots}
\usage{
posterior_predictive(bpc_object, n = 100)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{n}{number of times to sample from the posterior}
}
\value{
a list containing two values: y value that represents the collected data and y_pred that represents the predictive posterior matrix
}
\description{
Calculate the posterior predictive distributions.
Helps to check the fitness of the model. Use it in conjunction with the shinystan app to visualize some nice posterior predictive plots
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
pp<-posterior_predictive(m)
print(pp$y)
print(pp$y_pred)
}
}
