% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{print.bpc}
\alias{print.bpc}
\title{Print method for the bpc object}
\usage{
\method{print}{bpc}(x)
}
\arguments{
\item{x}{a bpc object}

\item{...}{}
}
\value{

}
\description{
Print method for the bpc object
}
