% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{rank_parameters}
\alias{rank_parameters}
\title{Generate a ranking of the parameter strength only and based on sampling the posterior distribution}
\usage{
rank_parameters(bpc_object, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{n}{Number of times we will sample the posterior}
}
\value{
a rank data frame
}
\description{
Generate a ranking of the parameter strength only and based on sampling the posterior distribution
}
