% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{rank_players}
\alias{rank_players}
\title{Generate a ranking of the ability based on sampling the posterior distribution of the ranks
To print this object you should remove the last column PosteriorRank since it contain the whole posterior distribution for each case}
\usage{
rank_players(bpc_object, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{n}{Number of times we will sample the posterior}
}
\value{
a data frame. This data frame contains the median of the rank, the mean, the standard deviation and column with a list containing all the posterior values for the rank
}
\description{
Generate a ranking of the ability based on sampling the posterior distribution of the ranks
To print this object you should remove the last column PosteriorRank since it contain the whole posterior distribution for each case
}
\examples{
m<-bpc(data = tennis_agresti, player0 = 'player0',player1 = 'player1',result_column = 'y',model_type = 'bt', solve_ties = 'none')
rank_m<-rank_parameters(m,n=100)
rank_table <- dplyr::select(rank_m,-MeanRank, -StadRank,-PosteriorRank)
print(rank_table)
}
