% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{replace_parameter_index_with_names}
\alias{replace_parameter_index_with_names}
\title{Replace the name of the parameter from index to name using a lookup_table
Receives a data frame and returns a dataframe
lambda\link{1} --> lambda_Biometrika}
\usage{
replace_parameter_index_with_names(data, column, par, lookup_table)
}
\arguments{
\item{data}{dataframe}

\item{column}{name of the colum}

\item{par}{name of the parameter}

\item{lookup_table}{a data frame}
}
\value{

}
\description{
Replace the name of the parameter from index to name using a lookup_table
Receives a data frame and returns a dataframe
lambda\link{1} --> lambda_Biometrika
}
