% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{sample_stanfit}
\alias{sample_stanfit}
\title{Return a data frame by resampling the posterior from a stanfit}
\usage{
sample_stanfit(stanfit, par, n = 100)
}
\arguments{
\item{stanfit}{stanfit object}

\item{par}{parameter name}

\item{n}{number of samples}
}
\value{
a dataframe containing the samples of the parameter. Each column is a parameter (in order of the index), each row is a sample
}
\description{
Return a data frame by resampling the posterior from a stanfit
}
