% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{save_bpc_model}
\alias{save_bpc_model}
\title{Thin wrapper to save the bpc model for examining later}
\usage{
save_bpc_model(bpc_object, filename, path = ".")
}
\arguments{
\item{bpc_object}{a bpc object}

\item{filename}{string with the file name}

\item{path}{string with path following the conventions of the operating system}
}
\description{
Thin wrapper to save the bpc model for examining later
}
