% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{summary.bpc}
\alias{summary.bpc}
\title{Summary of the model bpc model.}
\usage{
\method{summary}{bpc}(object, digits = 2, show_probabilities = TRUE, ...)
}
\arguments{
\item{object}{bpc object}

\item{digits}{number of decimal digits in the table}

\item{show_probabilities}{should the tables of probabilities (Table 2) be displayed or not. Default to T but it is recommended to turn to F if either it has a large number of players (15+) or a large number of players and cluster, the table grows combinatorially.}

\item{\dots}{additional parameters for the generic summary function. Not used.}
}
\description{
\itemize{
\item Table 1: Contains the parameter estimates and respective HPD interval
\item Table 2: Contains the posterior probability for the combination of all players
\item Table 3: Contains the ranking of the players' abilities based on the posterior distribution of the ranks
}
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
summary(m)
}
}
