% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_hpd.R
\name{summary_from_fit}
\alias{summary_from_fit}
\title{Summary for all parameters from a stanfit object
Here we use the coda package}
\usage{
summary_from_fit(fit, pars, credMass = 0.95)
}
\arguments{
\item{fit}{a cmdstanr fit object retrieved from a bpc object}

\item{pars}{the model parameters}

\item{credMass}{Credibility mass for the interval}
}
\value{
a data frame with the HPDI calculated from the coda package
}
\description{
Summary for all parameters from a stanfit object
Here we use the coda package
}
\references{
Martyn Plummer, Nicky Best, Kate Cowles and Karen Vines (2006). CODA: Convergence Diagnosis and Output Analysis for MCMC, R News, vol 6, 7-11
}
