% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis_agresti}
\alias{tennis_agresti}
\title{This is the expansion of the tennis data from  Agresti (2003) p.449
This data refers to matches for several women tennis players during 1989 and 1990}
\format{
This is the expansion of the data where each row contains 1 match only
\itemize{
\item player0: name of player0
\item player1: name of player1
\item y: corresponds to the result of the match: 0 if player0 won, 1 if player1 won.
\item id: is a column to make each row unique in the data. It does not have any particular interpretation
}
}
\source{
Agresti, Alan. Categorical data analysis. Vol. 482. John Wiley & Sons, 2003.
}
\usage{
tennis_agresti
}
\description{
This is the expansion of the tennis data from  Agresti (2003) p.449
This data refers to matches for several women tennis players during 1989 and 1990
}
\keyword{data}
