% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc.R
\name{bpc}
\alias{bpc}
\title{Bayesian Paired comparison  regression models in Stan}
\usage{
bpc(
  data,
  player0,
  player1,
  player0_score = NULL,
  player1_score = NULL,
  result_column = NULL,
  model_type = "bradleyterry",
  solve_ties = "random",
  win_score = "higher",
  prior_lambda_std = 2,
  chains = 4,
  iter = 2000,
  warmup = 1000
)
}
\arguments{
\item{data}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{player0}{A string with name of the column containing the players 0}

\item{player1}{A string with name of the column containing the players 0}

\item{player0_score}{A string with name of the column containing the scores of players 0}

\item{player1_score}{A string with name of the column containing the scores of players 1}

\item{result_column}{A string with name of the column containing the winners. 0 for player 0, 1 for player 1 and -1 for ties}

\item{model_type}{'bradleyterry' (default), 'davidson' for ties}

\item{solve_ties}{A string for the method of handling ties. 'random' for converting ties randomly, 'remove' for removing the tie occurrencess
'none' to ignore ties. The last requires a model to handle ties such as the davidson}

\item{win_score}{A string that indicates if a 'higher' score is winner or if a 'lower' score is winner}

\item{prior_lambda_std}{Positive value for the standard deviation of the normal prior if not uniform. Default is 2}

\item{chains}{Number of chains passed to Stan sampling. Positive integer, default=4. For more information consult Stan}

\item{iter}{Number of iterations passed to Stan sampling. Positive integer, default =2000. For more information consult Stan}

\item{warmup}{Number of iteration for the warmup passed to Stan sampling. Positive integer, default 1000.  For more information consult Stan}
}
\value{
An object of the class TODO:finish
}
\description{
Bayesian Paired comparison  regression models in Stan
}
\examples{
 bpc(data=citations_agresti,
              player0 = 'journal1',
              player1 = 'journal2',
              player0_score = 'score1',
              player1_score = 'score2',
              model_type='bradleyterry',
              solve_ties='random',
              win_score = 'higher')
}
