% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brasil_soccer_league}
\alias{brasil_soccer_league}
\title{This is a dataset  with the results matches fromo  the first league of the Brazilian soccer championship from 2017-2019.
It was reduced and translatedfrom the adaduque/Brasileirao_Dataset repository}
\format{
Data frame that contains 1140 matches and 9 Columns from the Brazilian soccer championship
\itemize{
\item Time: time of the day in 24h format
\item DayWeek: day of the week
\item Date: date YY-MM-DD
\item HomeTeam: name of the team playing home
\item VisitorTeam: name of the team playing visitor
\item Round: Round number of the championship
\item Stadium: Name of the stadium where the game was played
\item ScoreHomeTeam: number of goals for the home team
\item ScoreVisitorTeam: number of goals for the visitor
}
}
\source{
\url{https://github.com/adaoduque/Brasileirao_Dataset}
}
\usage{
brasil_soccer_league
}
\description{
This is a dataset  with the results matches fromo  the first league of the Brazilian soccer championship from 2017-2019.
It was reduced and translatedfrom the adaduque/Brasileirao_Dataset repository
}
\keyword{data}
