% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_checks.R
\name{check_if_there_are_na}
\alias{check_if_there_are_na}
\title{Check for NA in the specfic columns and returns T or F is there is at least 1 NA in those columns}
\usage{
check_if_there_are_na(
  d,
  player0,
  player1,
  player0_score = NULL,
  player1_score = NULL,
  result_column = NULL
)
}
\arguments{
\item{d}{a data frame}

\item{player0}{the name of column for player0}

\item{player1}{the name of column for player1}

\item{player0_score}{the name of column for player0 scores}

\item{player1_score}{the name of column for player1 scores}

\item{result_column}{the name of column for results}
}
\value{
TRUE (there are NA) or FALSE (no NA)
}
\description{
Check for NA in the specfic columns and returns T or F is there is at least 1 NA in those columns
}
