% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_checks.R
\name{check_if_there_are_ties}
\alias{check_if_there_are_ties}
\title{Check if a data frame column contains ties}
\usage{
check_if_there_are_ties(d_column)
}
\arguments{
\item{d_column}{a column with the values for the ties}
}
\value{
T (there are ties) or F (no ties)
}
\description{
Check if a data frame column contains ties
}
