% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_checks.R
\name{check_result_column}
\alias{check_result_column}
\title{Check if a data frame column contains only the values 1 0 and 2. Used to check the format of the results}
\usage{
check_result_column(d_column)
}
\arguments{
\item{d_column}{a column from a data frame}
}
\value{
TRUE (correct) or FALSE (with problems)
}
\description{
Check if a data frame column contains only the values 1 0 and 2. Used to check the format of the results
}
