% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers.R
\name{create_array_of_par_names}
\alias{create_array_of_par_names}
\title{Helper function for the get_par_names function in exports. This creates an array with the name of the parameter and the corresponding index}
\usage{
create_array_of_par_names(par, lookup_table)
}
\arguments{
\item{par}{a name for the parameter}

\item{lookup_table}{a lookuptable to convert the indexes to that}
}
\value{
a vector
}
\description{
Helper function for the get_par_names function in exports. This creates an array with the name of the parameter and the corresponding index
}
