% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_object.R
\name{create_bpc_object}
\alias{create_bpc_object}
\title{Creates a bpc object based on a stanfit object and a lookup_table dataframe}
\usage{
create_bpc_object(
  stanfit,
  lookup_table,
  cluster_lookup_table = NULL,
  model_type,
  standata,
  call_arg
)
}
\arguments{
\item{stanfit}{Stanfit object returned by rstan::sampling}

\item{lookup_table}{lookup_table dataframe. Two columns one Index the other Names where each each index will match a string in the names}
}
\value{
a bpc object
}
\description{
Creates a bpc object based on a stanfit object and a lookup_table dataframe
}
