% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{create_index_lookuptable}
\alias{create_index_lookuptable}
\title{Create a lookup table of names and indexes
Note that the indexes will be created in the order they appear. For string this doesnt make much difference but for numbers the index might be different than the actual number that appears in names}
\usage{
create_index_lookuptable(d, player0, player1)
}
\arguments{
\item{d}{A data frame containing the observations. The other parameters specify the name of the columns}

\item{player0}{The name of the column of data  contains player0}

\item{player1}{The name of the column of data  contains player0}
}
\value{
A dataframe of a lookup table with columns Names and Index
}
\description{
Create a lookup table of names and indexes
Note that the indexes will be created in the order they appear. For string this doesnt make much difference but for numbers the index might be different than the actual number that appears in names
}
