% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_loo}
\alias{get_loo}
\title{Tiny wrapper for the PSIS-LOO-CV method from the loo package.}
\usage{
get_loo(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\value{
a loo object
}
\description{
This is used to evaluate the fit of the model using entropy criteria
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
 l<-get_loo(m)
print(l)
}
}
\references{
Vehtari A, Gelman A, Gabry J (2017). Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC. Statistics and Computing_, 27, 1413-1432
}
