% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Return the mean and the HPDI of the parameters of the model}
\usage{
get_parameters(
  bpc_object,
  params = NULL,
  HPDI = TRUE,
  n_eff = TRUE,
  Rhat = FALSE
)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{params}{a vector with the parameters to use. If null then all will be present}

\item{HPDI}{should return the HPDI or the credible intervals. Default is returning the HPDI}

\item{n_eff}{Should include the number of effective samples in the df}

\item{Rhat}{Should include the Rhat in the df}
}
\value{
a data frame containing a column with the parameters, a column with mean and two columns with higher and lower intervals
}
\description{
Return a data frame with the mean and with high and low 95\% hpd interval or credible interval for all parameters of the model
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
hpdi<-get_parameters(m)
print(hpdi)}
}
