% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_get_probabilities.R
\name{get_probabilities_df}
\alias{get_probabilities_df}
\title{Get the empirical win/draw probabilities based on the ability/strength parameters.}
\usage{
get_probabilities_df(bpc_object, newdata = NULL, n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{newdata}{default to NULL. If used,  it will calculate the probabilities only for the newdata. Otherwise it will calculate for all combinations}

\item{n}{number of samples to draw from the posterior}
}
\value{
a data frame table with the respective probabilities
}
\description{
Get the empirical win/draw probabilities based on the ability/strength parameters.
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
prob<-get_probabilities_df(m)
print(prob)

# Now we can try just specifying a few data combinations of probabilities
m2<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
prob<-get_probabilities_df(m2, newdata=tennis_agresti[c(1,15),])
print(prob)

}
}
