% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_publication.R
\name{get_probabilities_table}
\alias{get_probabilities_table}
\title{Publication-ready table for the probabilities}
\usage{
get_probabilities_table(
  bpc_object,
  format = "latex",
  digits = 3,
  caption = "Estimated posterior probabilites"
)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{format}{A character string. same formats utilized in the knitr::kable function
\itemize{
\item 'latex': output in latex format
\item 'simple': appropriated for the console
\item 'pipe': Pandoc's pipe tables
\item 'html': for html formats
\item 'rst'
Possible values are latex, html, p, simple (Pandoc's simple tables), and rst.
}}

\item{digits}{number of digits in the table}

\item{caption}{a string containing the caption of the table}
}
\value{
a formatted table
}
\description{
Publication-ready table for the probabilities
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
t<-get_probabilities_table(m)
print(t)
}
}
