% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_stanfit}
\alias{get_stanfit}
\title{Retrieve the stanfit object}
\usage{
get_stanfit(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\value{
stanfit object
}
\description{
Retrieve the stanfit object
}
