% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{get_stanfit}
\alias{get_stanfit}
\title{Retrieve the stanfit object generated by rstan.}
\usage{
get_stanfit(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\value{
a stanfit object
}
\description{
This object can be used with any other function or package that uses stanfit objects from rstan
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
stanfit<- get_stanfit(m)
print(class(stanfit))
}
}
