% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_hpd.R
\name{HPDI_from_stanfit}
\alias{HPDI_from_stanfit}
\title{Calculate HPDI for all parameters from a stanfit object
Here we use the coda package}
\usage{
HPDI_from_stanfit(stanfit)
}
\arguments{
\item{stanfit}{a stanfit object retrived from a bpc object}
}
\value{
a data frame with the HPDI calculated from the coda package
}
\description{
Calculate HPDI for all parameters from a stanfit object
Here we use the coda package
}
\references{
Martyn Plummer, Nicky Best, Kate Cowles and Karen Vines (2006). CODA: Convergence Diagnosis and Output Analysis for MCMC, R News, vol 6, 7-11
}
