% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{is a real -inf to inf}
}
\value{
a value between 0 and 1
}
\description{
Inverse logit function
}
\examples{
inv_logit(5)
inv_logit(-5)
inv_logit(0)
}
\references{
https://en.wikipedia.org/wiki/Logit
}
