% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{launch_shinystan}
\alias{launch_shinystan}
\title{Tiny wrapper to launch a shinystan app to investigate the MCMC.}
\usage{
launch_shinystan(bpc_object)
}
\arguments{
\item{bpc_object}{a bpc object}
}
\description{
It launches a shinystan app automatically in the web browser
}
\examples{
\donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
launch_shinystan(m)
}
}
