% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{p is a probability 0 to 1}
}
\value{
a value between -inf and inf
}
\description{
Logit function
}
\examples{
logit(0.5)
logit(0.2)
}
\references{
https://en.wikipedia.org/wiki/Logit
}
