% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{predict.bpc}
\alias{predict.bpc}
\title{Title}
\usage{
\method{predict}{bpc}(object, newdata, n = 100)
}
\arguments{
\item{object}{a bpc object}

\item{newdata}{a data frame that contains columns with the same names as used to fit the data in the model.}

\item{n}{number of time we will iterate and get the posterior. default is 100 so we dont get too many}
}
\value{

}
\description{
Title
}
