% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{predict.bpc}
\alias{predict.bpc}
\title{Predict results for new data
This S3 function receives the bpc model and a data frame containing the same columns as the one used to fit the model.
It returns another data frame with with the same columns of the new data and n additional columns representing a posterior preditive distribution.}
\usage{
\method{predict}{bpc}(object, newdata, n = 100, return_matrix = F)
}
\arguments{
\item{object}{a bpc object}

\item{newdata}{a data frame that contains columns with the same names as used to fit the data in the model.}

\item{n}{number of time we will iterate and get the posterior. default is 100 so we dont get too many}

\item{return_matrix}{should we return only a matrix with the predictive values. Default F. Use this to combine with predictive posterior plots in bayesplot}
}
\value{
a dataframe or a matrix depending on the return_matrix parameter
}
\description{
Predict results for new data
This S3 function receives the bpc model and a data frame containing the same columns as the one used to fit the model.
It returns another data frame with with the same columns of the new data and n additional columns representing a posterior preditive distribution.
}
