% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{print.bpc}
\alias{print.bpc}
\title{Print method for the bpc object
We only print the parameter values and the HDPI}
\usage{
\method{print}{bpc}(x)
}
\arguments{
\item{x}{a bpc object}

\item{...}{additional parameters for the generic print function}
}
\value{

}
\description{
Print method for the bpc object
We only print the parameter values and the HDPI
}
