% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_helpers_indexes.R
\name{replace_parameter_index_with_names}
\alias{replace_parameter_index_with_names}
\title{Replace the name of the parameter from index to name using a lookup_table
Receives a data frame and returns a dataframe.
E.g  for parameter lambda with 4 players we have (lambda\link{1}, lambda\link{2}, lambda\link{3}, lambda\link{4}). This function will return (lambda\link{Biometrika}, lambda\link{CommStat}, lambda\link{JASA}, lambda\link{JRSSB})}
\usage{
replace_parameter_index_with_names(
  d,
  column,
  par,
  lookup_table,
  cluster_lookup_table = NULL,
  predictors_lookup_table = NULL
)
}
\arguments{
\item{d}{dataframe}

\item{column}{name of the colum}

\item{par}{name of the parameter}

\item{lookup_table}{lookup table of the players}

\item{cluster_lookup_table}{a lookup table of the predictors}
}
\value{
a data. frame where we change the names in the variable colum to the corresponding parameter_name from the lookup table
}
\description{
Replace the name of the parameter from index to name using a lookup_table
Receives a data frame and returns a dataframe.
E.g  for parameter lambda with 4 players we have (lambda\link{1}, lambda\link{2}, lambda\link{3}, lambda\link{4}). This function will return (lambda\link{Biometrika}, lambda\link{CommStat}, lambda\link{JASA}, lambda\link{JRSSB})
}
