% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_exports.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Get the posterior samples for the  parameters}
\usage{
sample_posterior(bpc_object, par = "lambda", n = 1000)
}
\arguments{
\item{bpc_object}{a bpc object}

\item{par}{name of the parameters to predict}

\item{n}{how many times are we sampling. Default 1000}
}
\value{
Return the posterior samples for the lambda parameters
}
\description{
Get the posterior samples for the  parameters
}
