% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{summary.bpc}
\alias{summary.bpc}
\title{Summary of the model bpc model
Contains only the parameters estimates and the respective HPD interval}
\usage{
\method{summary}{bpc}(x)
}
\arguments{
\item{x}{bpc object}

\item{...}{}
}
\value{

}
\description{
Summary of the model bpc model
Contains only the parameters estimates and the respective HPD interval
}
