% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpc_s3.R
\name{summary.bpc}
\alias{summary.bpc}
\title{Summary of the model bpc model
Table 1: Contains the parameter estimates and respective HPD interval
Table 2: Contains the empirical probability for the combination of all players}
\usage{
\method{summary}{bpc}(x)
}
\arguments{
\item{x}{bpc object}

\item{...}{additional parameters for the generic summary function}
}
\value{

}
\description{
Summary of the model bpc model
Table 1: Contains the parameter estimates and respective HPD interval
Table 2: Contains the empirical probability for the combination of all players
}
\examples{
m<-bpc(data = tennis_agresti, player0 = 'player0',player1 = 'player1',result_column = 'y',model_type = 'bt', solve_ties = 'none')
summary(m)
}
