% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{tennis_aggresti}
\alias{tennis_aggresti}
\title{This is the tennis data from  Agresti (2003) p.449
This data refers to matches for several women tennis players during 1989 and 1990
This is the aggregated form originally presented}
\description{
This is the tennis data from  Agresti (2003) p.449
This data refers to matches for several women tennis players during 1989 and 1990
This is the aggregated form originally presented
}
\references{
Agresti, Alan. Categorical data analysis. Vol. 482. John Wiley & Sons, 2003.
}
\keyword{data}
